+++ src/scalecast/Forecaster.py
@@ -1901,7 +1901,12 @@ class Forecaster:
             descriptive_assert(
                 len(self.current_xreg[k]) == len(self.y),
                 ForecastError,
-                "something went wrong when setting covariate values--try resetting the object and trying again",
+                "length of array representing the input '{}' does match the length of the stored observed values: ({} vs. {})."
+                " if you do not know how this happened, consider raising an issue: https://github.com/mikekeith52/scalecast/issues/new.".format(
+                    k,
+                    len(self.current_xreg[k]),
+                    len(self.y),
+                ),
             )
             self.future_xreg[k] = [float(x) for x in self.future_xreg[k]]
 
@@ -4250,13 +4255,29 @@ class Forecaster:
                 ]
         self.history[call_me]['CIPlusMinus'] = None
 
-    def reeval_cis(self,models='all'):
-        """ generates an expanding confidence interval that uses previously evaluated model classes to determine.
-        need to have evaluated at least three models to be able to use.
+    def reeval_cis(self, method = 'naive', models='all', n_iter=10, jump_back=1):
+        """ generates a dynamic confidence interval for passed models.
 
         Args:
+            method (str): one of "naive", "backtest".
+                if "naive", calculates confidence intervals by determining the standard deviation of each point of
+                every evaluated model passed to the function. time-steps that all models estimated closer together will receive smaller
+                intervals and steps where the models diverge significantly receive larger intervals.
+                this is a computationally cheap method but if all models perform similarly poorly, it can generate tight intervals, which is 
+                a downside. it is recommended to use at least 3 models with this method.
+                if "backtest", calculates confidence intervals by backtesting each model and determining the standard deviation of the out-of-sample
+                residual of each forecast step for both test-set predictions and actuals.
+                this is a generalizable way to obtain trustworthy confidence intervals for all models, but it is more computationally expensive,
+                it does not work for objects that were transformed and then reverted. it also does not work for combo models.
+                it is recommended to set n_iter to at least 3 with this method.
+                see https://scalecast.readthedocs.io/en/latest/Forecaster/Forecaster.html#src.scalecast.Forecaster.Forecaster.backtest.
             models (str or list-like): default 'all'. the models to regenerate cis for. 
                 needs to have at least 3 to work with.
+            n_iter (int) - default 10. the number of iterations to backtest. 
+                models will iteratively train on all data before the fcst_length worth of values. 
+                each iteration takes observations (this number is determined by the value passed to the jump_back arg) 
+                off the end to redo the cast until all of n_iter is exhausted. ignored when method == 'naive'.
+            jump_back (int) - default 1. the number of time steps between two consecutive training sets. ignored when method == 'naive'.
         Returns:
             None
 
@@ -4291,23 +4312,66 @@ class Forecaster:
             "LevelTSLowerCI": "LevelTestSetPreds",
         }
 
-        for m in models:
-            for i, kv in enumerate(attr_set_map.items()):
-                if i % 2 == 0:
-                    fcsts = np.array(
-                        [self.history[m][kv[1]] for m in models]
+        if method == 'naive':
+            if len(models) < 3:
+                warnings.warn('reeval_cis(method="naive") does not work well with fewer than three models.')
+            for m in models:
+                for i, kv in enumerate(attr_set_map.items()):
+                    if i % 2 == 0:
+                        fcsts = np.array(
+                            [self.history[m][kv[1]] for m in models]
+                        )
+                        self.history[m][kv[0]] = [
+                            self.history[m][kv[1]][idx] + set_ci_step(fcsts.std(axis=0)[idx],)
+                            for idx, fcst_step in enumerate(fcsts.mean(axis=0))
+                        ]
+                    else:
+                        self.history[m][kv[0]] = [
+                            self.history[m][kv[1]][idx] - set_ci_step(fcsts.std(axis=0)[idx],)
+                            for idx, fcst_step in enumerate(fcsts.mean(axis=0))
+                        ]
+        elif method == 'backtest':
+            for m in models:
+                bt_length = max(len(self.history[m]['Forecast']),self.history[m]['TestSetLength'])
+                try:
+                    self.backtest(model=m, fcst_length=bt_length, n_iter=n_iter, jump_back=jump_back)
+                except ForecastError:
+                    raise ForecastError(
+                        'backtested confidence intervals do not work when object was transformed and then reverted.'
+                        ' if object was not transformed and then reverted, raise an issue: https://github.com/mikekeith52/scalecast/issues/new.'
                     )
-                    self.history[m][kv[0]] = [
-                        self.history[m][kv[1]][idx] + set_ci_step(fcsts.std(axis=0)[idx],)
-                        for idx, fcst_step in enumerate(fcsts.mean(axis=0))
-                    ]
-                else:
-                    self.history[m][kv[0]] = [
-                        self.history[m][kv[1]][idx] - set_ci_step(fcsts.std(axis=0)[idx],)
-                        for idx, fcst_step in enumerate(fcsts.mean(axis=0))
-                    ]
-
+                results = self.export_backtest_values(m)
+                resids = pd.DataFrame(columns = ['resids','stepnum'])
+                for i, col in enumerate(results):
+                    if i%3 == 0:
+                        resids = pd.concat(
+                            [
+                                resids,
+                                pd.DataFrame(
+                                    {
+                                        'resids':(results.iloc[:,i+1] - results.iloc[:,i+2]).to_list(),
+                                        'stepnum':[i//3]*bt_length
+                                    }
+                                )
+                            ]
+                        )
+                resids_std = resids.std(axis=1)
+                for i, kv in enumerate(attr_set_map.items()):
+                    vals = self.history[m][kv[1]]
+                    resids_tmp = resids_std.iloc[:len(vals)]
+                    if i % 2 == 0:
+                        self.history[m][kv[0]] = [
+                            vals[idx] + set_ci_step(val)
+                            for idx, val in enumerate(resids_tmp)
+                        ]
+                    else:
+                        self.history[m][kv[0]] = [
+                            vals[idx] - set_ci_step(val)
+                            for idx, val in enumerate(resids_tmp)
+                        ]
             self.history[m]['CIPlusMinus'] = None
+        else:
+            raise ValueError(f'method expected one of "naive", "backtest", got {method}')
 
     def add_sklearn_estimator(self, imported_module, called):
         """ adds a new estimator from scikit-learn not built-in to the forecaster object that can be called using set_estimator().
@@ -4881,6 +4945,52 @@ class Forecaster:
                     alpha=0.2,
                     color=_colors_[i],
                     label="{} {:.0%} CI".format(m, self.history[m]["CILevel"]),
+                )
+
+        plt.legend()
+        plt.xlabel("Date")
+        plt.ylabel("Values")
+        return ax
+
+    def plot_backtest_values(self,model,figsize=(12,6)):
+        """ plots all backtest values over every iteration. only plots one model at a time.
+
+        Args:
+            model (str): the model nickname to plot the backtest values for. must have called f.backtest(model)
+                previously.
+            figsize (tuple): default (12,6). size of the resulting figure.
+        
+        Returns:
+            (Axis): the figure's axis.
+
+        >>> f.set_estimator('elasticnet')
+        >>> f.manual_forecast(alpha=.2)
+        >>> f.backtest('elasticnet')
+        >>> f.plot_backtest_values('elasticnet') # plots all backtest values
+        >>> plt.show()
+        """
+        _, ax = plt.subplots(figsize=figsize)
+        values = self.export_backtest_values(model)
+        y = self.levely[:]
+        dates = self.current_dates.to_list()
+        ac_len = min(len(y),len(dates))
+
+        sns.lineplot(
+            x = dates[-ac_len:],
+            y = y[-ac_len:],
+            label="actuals",
+            ax=ax,
+        )
+
+        for i, col in enumerate(values):
+            if i % 3 == 0:
+                sns.lineplot(
+                    x = values.iloc[:,i], # dates
+                    y = values.iloc[:,i+2], # predictions
+                    label = f'iter {i//3+1}',
+                    ax = ax,
+                    color=_colors_[i//3],
+                    alpha = 0.7,
                 )
 
         plt.legend()
@@ -5522,7 +5632,7 @@ class Forecaster:
         Args:
             model (str): the model to run the backtest for. use the model nickname.
             fcst_length (int or str): default 'auto'. 
-                if 'auto', uses the same forecast length as saved in the object currently.
+                if 'auto', uses the same forecast length as the number of future dates saved in the object currently.
                 if int, uses that as the forecast length.
             n_iter (int): default 10. the number of iterations to backtest.
                 models will iteratively train on all data before the fcst_length worth of values.
@@ -5585,6 +5695,7 @@ class Forecaster:
             metric_results.loc["MAPE", f"iter{i+1}"] = test_mets[
                 "LevelTestSetMAPE"
             ].values[0]
+            value_results[f"iter{i+1}dates"] = test_preds['DATE'].values.copy()
             value_results[f"iter{i+1}actuals"] = test_preds.iloc[:, 1].values.copy()
             value_results[f"iter{i+1}preds"] = test_preds.iloc[:, 2].values.copy()
 
@@ -5607,6 +5718,8 @@ class Forecaster:
     def export_backtest_values(self, model):
         """ extracts the backtest values for a given model.
         only works if `backtest()` has been called.
+        the DataFrame will return columns for the date (first), actuals (second), and predictions (third)
+        across every backtest iteration (10 by default).
 
         Args:
             model (str): the model nickname to extract values for.