+++ docs/change_log.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project are documented in this file since v0.1.8. The source code for most releases is available on [GitHub](https://github.com/mikekeith52/scalecast).
 
+## [0.15.8] - 2022-11-02
+### Added
+- added `method` argument to `Forecaster.reeval_cis()` and `MVForecaster.reeval_cis()` and included an option to get confidence intervals through backtesting.
+- added `Forcaster.plot_backtest_values()`.
+### Changed
+- changed some error messages to be more descriptive and to encourage raising issues on github.
+- changed how level confidence intervals are obtained when calling `SeriesTransformer.DiffRevert()` to be more efficient.
+- the backtest_values DataFrame now includes dates in `Forecaster` and `MVForecaster`. the order of the dataframe is Date --> Actuals --> Preds for all iterations.
+### Fixed
+
 ## [0.15.7] - 2022-10-28
 ### Added
 ### Changed